chore: Split e2e test cases
