+++ .github/workflows/CI.yml
@@ -91,8 +91,8 @@ jobs:
         with:
           name: coverage-out
           path: coverage.out
-  e2e:
-    name: e2e
+  e2e-cli:
+    name: e2e-cli
     strategy:
       matrix:
         os: [ubuntu-latest]
@@ -116,15 +116,15 @@ jobs:
       - uses: actions-ecosystem/action-get-latest-tag@v1
         id: get-latest-tag
       - name: e2e test
-        run: make e2e-test
+        run: make e2e-cli-test
         env:
           GIT_LATEST_TAG: ${{ steps.get-latest-tag.outputs.tag }}
       - name: Upload coverage report
         uses: actions/upload-artifact@v3
         with:
-          name: e2e-coverage-out
-          path: e2e-coverage.out
-  e2e-doc-test:
+          name: e2e-cli-coverage-out
+          path: e2e-cli-coverage.out
+  e2e-doc:
     name: e2e-doc
     strategy:
       matrix:
@@ -147,7 +147,7 @@ jobs:
             ${{ runner.OS }}-build-
             ${{ runner.OS }}-
       - uses: actions-ecosystem/action-get-latest-tag@v1
-        id: get-latest-tag    
+        id: get-latest-tag
       - name: e2e doc test
         run: make e2e-doc-test
         env:
@@ -157,6 +157,39 @@ jobs:
         with:
           name: e2e-doc-coverage-out
           path: e2e-doc-coverage.out
+  e2e-lang:
+    name: e2e-lang
+    strategy:
+      matrix:
+        os: [ubuntu-latest]
+    runs-on: ${{ matrix.os }}
+    steps:
+      - name: Check out code
+        uses: actions/checkout@v2
+      - name: Setup Go
+        uses: actions/setup-go@v2
+        with:
+          go-version: 1.18
+      - name: Cache Go modules
+        uses: actions/cache@preview
+        with:
+          path: ~/go/pkg/mod
+          key: ${{ runner.os }}-build-${{ hashFiles('**/go.sum') }}
+          restore-keys: |
+            ${{ runner.OS }}-build-${{ env.cache-name }}-
+            ${{ runner.OS }}-build-
+            ${{ runner.OS }}-
+      - uses: actions-ecosystem/action-get-latest-tag@v1
+        id: get-latest-tag
+      - name: e2e test
+        run: make e2e-lang-test
+        env:
+          GIT_LATEST_TAG: ${{ steps.get-latest-tag.outputs.tag }}
+      - name: Upload coverage report
+        uses: actions/upload-artifact@v3
+        with:
+          name: e2e-lang-coverage-out
+          path: e2e-lang-coverage.out
   build:
     name: build
     strategy:
@@ -186,7 +219,8 @@ jobs:
   report:
     needs:
       - test
-      - e2e
+      - e2e-cli
+      - e2e-lang
     runs-on: ubuntu-latest
     steps:
       - name: Check out code
@@ -204,11 +238,16 @@ jobs:
         with:
           name: coverage-out
           path: coverage.out
-      - name: Get coverage report
+      - name: Get cli e2e coverage report
         uses: actions/download-artifact@v3
         with:
-          name: e2e-coverage-out
-          path: e2e-coverage.out
+          name: e2e-cli-coverage-out
+          path: e2e-cli-coverage.out
+      - name: Get language e2e coverage report
+        uses: actions/download-artifact@v3
+        with:
+          name: e2e-lang-coverage-out
+          path: e2e-lang-coverage.out
       # - name: Send coverage
       #   env:
       #     COVERALLS_TOKEN: ${{ secrets.GITHUB_TOKEN }}
