+++ e2e/cli/quick_start_test.go
@@ -12,17 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package e2e
+package cli
 
 import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
+	"github.com/tensorchord/envd/e2e"
 )
 
 var _ = Describe("e2e quickstart", Ordered, func() {
 	exampleName := "quick-start"
 	testcase := "e2e"
-	e := NewExample(exampleName, testcase)
+	e := e2e.NewExample(exampleName, testcase)
 	BeforeAll(e.BuildImage(true))
 	BeforeEach(e.RunContainer())
 	It("execute python demo.py", func() {
