+++ e2e/cli/testdata/quick-start/build.envd
