+++ e2e/cli/testdata/quick-start/demo.py
